% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_picture.R
\name{app_picture}
\alias{app_picture}
\title{Assemble the PICTURE App}
\usage{
app_picture(config_yaml = NULL, config = "default")
}
\value{
the PICTURE shiny app
}
\description{
Helper function to assemble the UI and server functions for the app
}
\seealso{
Other app_picture:
\code{\link{app_picture_run}()}
}
\concept{app_picture}
