% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_picture.R
\name{app_picture_run}
\alias{app_picture_run}
\title{Run PICTURE}
\usage{
app_picture_run(...)
}
\description{
Helper function to run PICTURE locally
Run this in the local folder as:
devtools::load_all(); picture.platform::app_picture_run()
As with shiny::runApp, this does not return until ctrl+c or closed UI
}
\seealso{
Other app_picture:
\code{\link{app_picture}()}
}
\concept{app_picture}
