% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_consts.R
\docType{data}
\name{deafult_start_date}
\alias{deafult_start_date}
\title{Default start date for selection}
\format{
An object of class \code{Date} of length 1.
}
\usage{
deafult_start_date
}
\description{
Default start date for selection
}
\keyword{datasets}
