% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_picture.R
\name{.app_yaml_to_config}
\alias{.app_yaml_to_config}
\title{Convert the app config yaml to an app config structure}
\usage{
.app_yaml_to_config(app_yamls)
}
\description{
Convert the app config yaml to an app config structure
}
