% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_picture.R
\name{.load_config_yaml}
\alias{.load_config_yaml}
\title{Load the selected config YAML and format inputs}
\usage{
.load_config_yaml(config_yaml, config)
}
\description{
Load the selected config YAML and format inputs
}
