% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cohort_formatting.R
\name{format_cohort_spec}
\alias{format_cohort_spec}
\title{Print a cohort description as text}
\usage{
format_cohort_spec(df_spec, idx)
}
\arguments{
\item{df_spec}{The full cohort specification}

\item{idx}{The index of the stage of the cohort to format}
}
\description{
Print a cohort description as text
}
