% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rdv_lookups.R
\name{get_descriptions_in}
\alias{get_descriptions_in}
\title{Get all the dataset description yaml files}
\usage{
get_descriptions_in(directory)
}
\description{
Get all the dataset description yaml files
}
