% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_logger.R
\name{get_packages_info}
\alias{get_packages_info}
\title{Wrapper for Sys.getenv}
\usage{
get_packages_info()
}
\description{
Description
}
\seealso{
Other utils_logger:
\code{\link{get_env_info}()},
\code{\link{get_system_info}()}
}
\concept{utils_logger}
