% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rdv_lookups.R
\name{get_rdv_cols}
\alias{get_rdv_cols}
\title{Function to call all functions needed to get a list of RDVs and columns
to load}
\usage{
get_rdv_cols(yml, all_rdvs)
}
\arguments{
\item{yml}{list; The app_yaml loaded as a list}

\item{all_rdvs}{dataframe; Dplyr dataframe of rdvs to load}
}
\value{
comb list; A set of RDVs and columns to load. RDVs are the names and
the columns are the values.
}
\description{
Function to call all functions needed to get a list of RDVs and columns
to load
}
\concept{limited}
