% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rdv_lookups.R
\name{get_rdvs_in}
\alias{get_rdvs_in}
\title{Function to extract all the valid-looking RDV CSVs in a directory}
\usage{
get_rdvs_in(directory)
}
\description{
Function to extract all the valid-looking RDV CSVs in a directory
}
