% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_logger.R
\name{get_system_info}
\alias{get_system_info}
\title{Provide R and OS version}
\usage{
get_system_info()
}
\description{
Description
}
\seealso{
Other utils_logger:
\code{\link{get_env_info}()},
\code{\link{get_packages_info}()}
}
\concept{utils_logger}
