% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_load_rdvs-shiny.R
\name{load_dataset_description}
\alias{load_dataset_description}
\title{Load the description file for the selected specialty, if available}
\usage{
load_dataset_description(data_dir, specialty)
}
\description{
Load the description file for the selected specialty, if available
}
