% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_consts.R
\docType{data}
\name{rdv_list}
\alias{rdv_list}
\title{List of RDVs and derived RDVs that are usable in PICTURE
It is currently necessary to hardcode these because of the way that reactives
work.}
\format{
An object of class \code{list} of length 11.
}
\usage{
rdv_list
}
\description{
List of RDVs and derived RDVs that are usable in PICTURE
It is currently necessary to hardcode these because of the way that reactives
work.
}
\keyword{datasets}
