% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_bootstrap_gosh-shiny.R
\name{ui_bootstrap_gosh}
\alias{ui_bootstrap_gosh}
\title{Core UI structure for bootstrap 4 based basic shiny UI}
\usage{
ui_bootstrap_gosh(logo_xl, logo_xs, sidebar_menu, body, ctrlbar = NULL)
}
\value{
Shiny taglist for the bootstrap based app UI
}
\description{
Core UI structure for bootstrap 4 based basic shiny UI
}
