% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_info_panel.R
\name{ui_info_panel}
\alias{ui_info_panel}
\title{Function to generate standardised info panel designs}
\usage{
ui_info_panel(input_filename)
}
\description{
Function to generate standardised info panel designs
}
