% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_load_rdvs-shiny.R
\name{ui_load_rdvs_server}
\alias{ui_load_rdvs_server}
\title{Helper (UI-less) server function to load all RDVs for a given specialty
It can load in either all rdvs in a folder or if "limited" is specified in app
yaml it will load only the rdvs and rdv columns specified in the app yaml}
\usage{
ui_load_rdvs_server(
  data_dir,
  select_specialty = NULL,
  n_max = Inf,
  selected_rdvs = NULL,
  app_yaml = NULL,
  show_progress = TRUE
)
}
\arguments{
\item{data_dir}{string; file path of rdvs to load}

\item{select_specialty}{string;  name of hospital specialty to use}

\item{n_max}{int; max rows to load}

\item{selected_rdvs}{list; list of rdvs and columns that are needed}

\item{app_yaml}{list; app yaml loaded as a list}

\item{show_progress}{boolean; show progress bar}
}
\value{
rdvs list; list of rdvs loaded into picture as tibbles
}
\description{
Helper (UI-less) server function to load all RDVs for a given specialty
It can load in either all rdvs in a folder or if "limited" is specified in app
yaml it will load only the rdvs and rdv columns specified in the app yaml
}
