% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_select_cohort-shiny.R
\name{ui_select_cohort_server}
\alias{ui_select_cohort_server}
\title{Shiny server module for the Filter a Patient Cohort}
\usage{
ui_select_cohort_server(rdvs_server, input_cohorts)
}
\arguments{
\item{xxxxxxx}{DESCRIPTION.}
}
\value{
Shiny moduleServer function for the server functionality
}
\description{
Shiny server function encapsulated by namespace 'id'
}
\seealso{
Other ui_select_cohort:
\code{\link{ui_select_cohort_ui}()}
}
\concept{ui_select_cohort}
