% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_select_cohort-shiny.R
\name{ui_select_cohort_ui}
\alias{ui_select_cohort_ui}
\title{Shiny UI module for the Filter a Patient Cohort}
\usage{
ui_select_cohort_ui()
}
\arguments{
\item{id}{Character label of the namespace for the module instance.}
}
\value{
Shiny taglist for the UI
}
\description{
Shiny UI function encapsulated by namespace 'id'
}
\seealso{
Other ui_select_cohort:
\code{\link{ui_select_cohort_server}()}
}
\concept{ui_select_cohort}
