% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_select_patient-shiny.R
\name{ui_select_patient_server}
\alias{ui_select_patient_server}
\title{Shiny server module for the Select a Patient ID}
\usage{
ui_select_patient_server(id, rdv_server, initial_cohorts)
}
\arguments{
\item{id}{Character label of the namespace for the module instance.}

\item{rdv_server}{DESCRIPTION.}
}
\value{
Shiny moduleServer function for the server functionality
}
\description{
Shiny server function encapsulated by namespace 'id'
}
\seealso{
Other ui_select_patient:
\code{\link{ui_select_patient_ui}()}
}
\concept{ui_select_patient}
