% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_select_patient-shiny.R
\name{ui_select_patient_ui}
\alias{ui_select_patient_ui}
\title{Shiny UI module for the Select a Patient ID}
\usage{
ui_select_patient_ui(id)
}
\arguments{
\item{id}{Character label of the namespace for the module instance.}
}
\value{
Shiny taglist for the UI
}
\description{
Shiny UI function encapsulated by namespace 'id'
}
\seealso{
Other ui_select_patient:
\code{\link{ui_select_patient_server}()}
}
\concept{ui_select_patient}
