% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_scaffold_new_module.R
\name{utils_scaffold_new_module}
\alias{utils_scaffold_new_module}
\title{Small function to help scaffold a new module for the package}
\usage{
utils_scaffold_new_module(
  name,
  title,
  inputs = NULL,
  target_dir = NULL,
  use_report = TRUE,
  use_testthat = TRUE,
  use_about = TRUE
)
}
\arguments{
\item{name}{R code name to use for module (e.g. my_module_analysis).
It is recommended to start the module with the three character
abbreviation of the primary RDV type (e.g. "dia_NAME" for
diagnoses), or "gen_NAME" in the case of a generic module that
can work across many RDVs.}

\item{title}{Pretty text name for the module (e.g. "Analysis of something")}

\item{inputs}{List of function input names, if known
(e.g. "df_pde, df_dia, some_constant"), optional}
}
\description{
Uses templates to scaffold the main files needed to create a new module
}
